+++ R/nonprob_dr.R
@@ -382,7 +382,7 @@ nonprob_dr <- function(selection,
                               pop_size_fixed = pop_size_fixed,
                               bias_corr_start = bias_corr_start)
 
-        var_total <- apply(boot_obj, 2, var)
+        var_total <- apply(boot_obj, 2, var, na.rm = TRUE)
         SE_values <- replicate(NROW(outcomes[[1]]), data.frame(nonprob = NA, prob = NA), simplify = FALSE)
         SE <- sqrt(var_total)
         output <- list(data.frame(mean = mu_hat, SE = SE))
@@ -639,7 +639,7 @@ nonprob_dr <- function(selection,
                             pop_size_fixed = pop_size_fixed,
                             bias_corr_start = bias_corr_start)
 
-        var_total <- apply(boot_obj, 2, var)
+        var_total <- apply(boot_obj, 2, var, na.rm = TRUE)
         SE_values <- replicate(NROW(outcomes[[1]]), data.frame(nonprob = NA, prob = NA), simplify = FALSE)
         SE <- sqrt(var_total)
         output <- list(data.frame(mean = mu_hat, SE = SE))